requirements.txt
# 财报造假分析系统依赖库

# 核心依赖
python>=3.8

# 数据处理和分析
pandas>=2.0.0
numpy>=1.24.0

# PDF处理（用于实际PDF解析）
pdfplumber>=0.10.3

# 配置管理
pyyaml>=6.0

# 命令行工具
click>=8.1.0

# 日志管理
loguru>=0.7.0

# 开发依赖（可选）
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0

# 使用说明：
# 1. 基础安装：pip install -r requirements.txt
# 2. 包含PDF解析：已包含pdfplumber
# 3. 开发工具：取消注释开发依赖并安装